+++ examples/bootstrap.rs
@@ -14,5 +14,5 @@ fn main() {
 
     let info = client.info().unwrap();
 
-    dbg!(info);
+    println!("{:?}", info);
 }